requirements.txt
typer
litellm
pydantic
pytest
pyyaml
